backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
gunicorn==25.3.0
requests==2.33.1
python-dotenv==1.2.2
